requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dill==0.3.8
Flask==3.0.3
ibm-cloud-sdk-core==3.20.3
ibm-platform-services==0.55.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==2.0.0
openqasm3==1.0.0
packaging==24.1
pbr==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyspnego==0.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qbraid==0.7.1
qbraid-core==0.1.17
qiskit==1.1.1
qiskit-ibm-runtime==0.25.0
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
scipy==1.14.0
setuptools==70.3.0
six==1.16.0
stevedore==5.2.0
symengine==0.11.0
sympy==1.13.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
